move state and functions in app.js and pass throught props
